requirements_dev.txt
# Generic python dependencies
black==25.1.0
mypy==1.15.0
pip==25.1.1
pre-commit==4.2.0
ruff==0.11.12

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.248
syrupy>=4.6.1

# Component dependencies
hassil>=2.0.4
home-assistant-intents>=2024.4.3

google-adk==1.0.0
aiofiles==24.1.0

# Eval dependencies
synthetic-home==4.6.3
home-assistant-datasets==0.2.0
